requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.1.1
gunicorn==23.0.0
idna==2.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mongoengine==0.27.0
mongomock==4.1.2
packaging==23.1
Pillow==10.1.0
pluggy==1.3.0
pymongo==4.6.0
pytest==7.4.1
pytest-flask==1.2.0
python-dotenv==0.10.5
requests==2.32.3
sentinels==1.0.0
six==1.16.0
soupsieve==2.5
tomli==2.0.1
urllib3==2.2.3
vercel==0.2.1
waitress==2.1.2
Werkzeug==2.2.3
WTForms==3.0.1
